my_agent/requirements.txt
google-adk==1.18.0
google-genai
pytz
python-dotenv
